# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScanInput {
  /**
     * Website URL to scan
     * @minLength 1
     */
  url: string;
}

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  analyzing: 'analyzing',
  completed: 'completed',
  failed: 'failed',
} as const;

export type AiRecommendationImpact = typeof AiRecommendationImpact[keyof typeof AiRecommendationImpact];


export const AiRecommendationImpact = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type AiRecommendationDifficulty = typeof AiRecommendationDifficulty[keyof typeof AiRecommendationDifficulty];


export const AiRecommendationDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type AiRecommendationConfidence = typeof AiRecommendationConfidence[keyof typeof AiRecommendationConfidence];


export const AiRecommendationConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface AiRecommendation {
  title: string;
  whatFound: string;
  whyImportant: string;
  whatToImprove: string;
  /** @nullable */
  proposal: string | null;
  impact: AiRecommendationImpact;
  difficulty: AiRecommendationDifficulty;
  confidence: AiRecommendationConfidence;
  basedOnChecks: string[];
}

export type CategoryScoreKey = typeof CategoryScoreKey[keyof typeof CategoryScoreKey];


export const CategoryScoreKey = {
  conversie: 'conversie',
  seo: 'seo',
  mobiel: 'mobiel',
  techniek: 'techniek',
  content: 'content',
  vertrouwen: 'vertrouwen',
  lokaal: 'lokaal',
} as const;

export type CategoryCheckStatus = typeof CategoryCheckStatus[keyof typeof CategoryCheckStatus];


export const CategoryCheckStatus = {
  passed: 'passed',
  failed: 'failed',
  unknown: 'unknown',
} as const;

export interface CategoryCheck {
  key: string;
  label: string;
  status: CategoryCheckStatus;
  /** @nullable */
  value: string | null;
  evidence: string;
  /**
     * @minimum 1
     * @maximum 3
     */
  weight: number;
}

export interface CategoryScore {
  key: CategoryScoreKey;
  label: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  checked: boolean;
  /** @minimum 0 */
  passedCount: number;
  /** @minimum 0 */
  failedCount: number;
  /** @minimum 0 */
  unknownCount: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  coveragePercent: number;
  note: string;
  checks: CategoryCheck[];
}

export interface DetectedFacts {
  /** @nullable */
  pageTitle: string | null;
  /** @minimum 0 */
  pageTitleLength: number;
  /** @nullable */
  metaDescription: string | null;
  /** @minimum 0 */
  metaDescriptionLength: number;
  /** @minimum 0 */
  h1Count: number;
  /** @minimum 0 */
  headingCount: number;
  headings: string[];
  /** @minimum 0 */
  visibleTextLength: number;
  /** @minimum 0 */
  linkCount: number;
  /** @minimum 0 */
  internalLinkCount: number;
  /** @minimum 0 */
  externalLinkCount: number;
  /** @minimum 0 */
  imageCount: number;
  /** @minimum 0 */
  imagesWithAlt: number;
  /** @minimum 0 */
  ctaCount: number;
  callsToAction: string[];
  /** @nullable */
  primaryCta: string | null;
  /** @nullable */
  primaryCtaClearlyMarked: boolean | null;
  /** @nullable */
  ctaAboveFold: boolean | null;
  contactSignals: string[];
  legalSignals: string[];
  companySignals: string[];
  socialProofSignals: string[];
  localSignals: string[];
  technicalSignals: string[];
  /** @minimum 0 */
  responseTimeMs: number;
  /**
     * @minimum 100
     * @maximum 599
     */
  httpStatus: number;
  /** @minimum 0 */
  pageSizeKb: number;
  https: boolean;
  /** @nullable */
  compressed: boolean | null;
  /** @nullable */
  contentEncoding: string | null;
  /** @nullable */
  canonical: string | null;
  robotsDirectives: string[];
  openGraphSignals: string[];
  hasViewport: boolean;
  hasLanguage: boolean;
  hasRobotsTxt: boolean;
  hasSitemap: boolean;
  /** @nullable */
  sitemapUrl: string | null;
  localBusinessStructuredData: boolean;
  mapsLink: boolean;
  placeSignal: boolean;
  regionSignal: boolean;
  /** @nullable */
  valuePropositionSignal: boolean | null;
  /** @nullable */
  targetAudienceSignal: boolean | null;
  /** @nullable */
  duplicateTextDetected: boolean | null;
}

export type ScanIssueSeverity = typeof ScanIssueSeverity[keyof typeof ScanIssueSeverity];


export const ScanIssueSeverity = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type ScanIssueImpact = typeof ScanIssueImpact[keyof typeof ScanIssueImpact];


export const ScanIssueImpact = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type ScanIssueDifficulty = typeof ScanIssueDifficulty[keyof typeof ScanIssueDifficulty];


export const ScanIssueDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface ScanIssue {
  id: string;
  title: string;
  severity: ScanIssueSeverity;
  impact: ScanIssueImpact;
  difficulty: ScanIssueDifficulty;
  fact: string;
  whyItMatters: string;
  recommendation: string;
  relatedChecks: string[];
}

export interface ScanAnalysis {
  /**
     * @minimum 0
     * @maximum 100
     */
  overallScore: number;
  aiRecommendations: AiRecommendation[] | null;
  categoryScores: CategoryScore[];
  detectedFacts: DetectedFacts;
  notChecked: string[];
  issues: ScanIssue[];
}

export interface Scan {
  /** @minimum 1 */
  id: number;
  url: string;
  status: ScanStatus;
  createdAt: string;
  analysis: ScanAnalysis | null;
  /** @nullable */
  error: string | null;
}

export interface ErrorResponse {
  error: string;
}