# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scans
    description: Website scan operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /scans:
    post:
      operationId: createScan
      tags: [scans]
      summary: Start a website scan
      description: Fetches the public homepage and creates a first analysis from the information that can actually be checked.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "201":
          description: Scan completed or marked as unable to complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "400":
          description: Invalid website URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listScans
      tags: [scans]
      summary: List recent website scans
      description: Returns the most recent scan requests and any available first analysis.
      responses:
        "200":
          description: Recent scan requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
  /scans/{scanId}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get a website scan result
      description: Returns the current lifecycle status and any available first analysis.
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Scan request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ScanInput:
      type: object
      properties:
        url:
          type: string
          format: uri
          minLength: 1
          description: Website URL to scan
      required:
        - url
    Scan:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        url:
          type: string
          format: uri
        status:
          type: string
          enum: [analyzing, completed, failed]
        createdAt:
          type: string
          format: date-time
        analysis:
          oneOf:
            - $ref: "#/components/schemas/ScanAnalysis"
            - type: "null"
        error:
          type: ["string", "null"]
      required:
        - id
        - url
        - status
        - createdAt
        - analysis
        - error
    ScanAnalysis:
      type: object
      properties:
        overallScore:
          type: integer
          minimum: 0
          maximum: 100
        aiRecommendations:
          oneOf:
            - type: array
              maxItems: 5
              items:
                $ref: "#/components/schemas/AiRecommendation"
            - type: "null"
        categoryScores:
          type: array
          items:
            $ref: "#/components/schemas/CategoryScore"
        detectedFacts:
          $ref: "#/components/schemas/DetectedFacts"
        notChecked:
          type: array
          items:
            type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ScanIssue"
      required:
        - overallScore
        - aiRecommendations
        - categoryScores
        - detectedFacts
        - notChecked
        - issues
    CategoryScore:
      type: object
      properties:
        key:
          type: string
          enum: [conversie, seo, mobiel, techniek, content, vertrouwen, lokaal]
        label:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        checked:
          type: boolean
        passedCount:
          type: integer
          minimum: 0
        failedCount:
          type: integer
          minimum: 0
        unknownCount:
          type: integer
          minimum: 0
        coveragePercent:
          type: integer
          minimum: 0
          maximum: 100
        note:
          type: string
        checks:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCheck"
      required:
        - key
        - label
        - score
        - checked
        - passedCount
        - failedCount
        - unknownCount
        - coveragePercent
        - note
        - checks
    CategoryCheck:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [passed, failed, unknown]
        value:
          type: ["string", "null"]
        evidence:
          type: string
        weight:
          type: integer
          minimum: 1
          maximum: 3
      required:
        - key
        - label
        - status
        - value
        - evidence
        - weight
    DetectedFacts:
      type: object
      properties:
        pageTitle:
          type: ["string", "null"]
        pageTitleLength:
          type: integer
          minimum: 0
        metaDescription:
          type: ["string", "null"]
        metaDescriptionLength:
          type: integer
          minimum: 0
        h1Count:
          type: integer
          minimum: 0
        headingCount:
          type: integer
          minimum: 0
        headings:
          type: array
          items:
            type: string
        visibleTextLength:
          type: integer
          minimum: 0
        linkCount:
          type: integer
          minimum: 0
        internalLinkCount:
          type: integer
          minimum: 0
        externalLinkCount:
          type: integer
          minimum: 0
        imageCount:
          type: integer
          minimum: 0
        imagesWithAlt:
          type: integer
          minimum: 0
        ctaCount:
          type: integer
          minimum: 0
        callsToAction:
          type: array
          items:
            type: string
        primaryCta:
          type: ["string", "null"]
        primaryCtaClearlyMarked:
          type: ["boolean", "null"]
        ctaAboveFold:
          type: ["boolean", "null"]
        contactSignals:
          type: array
          items:
            type: string
        legalSignals:
          type: array
          items:
            type: string
        companySignals:
          type: array
          items:
            type: string
        socialProofSignals:
          type: array
          items:
            type: string
        localSignals:
          type: array
          items:
            type: string
        technicalSignals:
          type: array
          items:
            type: string
        responseTimeMs:
          type: integer
          minimum: 0
        httpStatus:
          type: integer
          minimum: 100
          maximum: 599
        pageSizeKb:
          type: number
          minimum: 0
        https:
          type: boolean
        compressed:
          type: ["boolean", "null"]
        contentEncoding:
          type: ["string", "null"]
        canonical:
          type: ["string", "null"]
        robotsDirectives:
          type: array
          items:
            type: string
        openGraphSignals:
          type: array
          items:
            type: string
        hasViewport:
          type: boolean
        hasLanguage:
          type: boolean
        hasRobotsTxt:
          type: boolean
        hasSitemap:
          type: boolean
        sitemapUrl:
          type: ["string", "null"]
        localBusinessStructuredData:
          type: boolean
        mapsLink:
          type: boolean
        placeSignal:
          type: boolean
        regionSignal:
          type: boolean
        valuePropositionSignal:
          type: ["boolean", "null"]
        targetAudienceSignal:
          type: ["boolean", "null"]
        duplicateTextDetected:
          type: ["boolean", "null"]
      required:
        - pageTitle
        - pageTitleLength
        - metaDescription
        - metaDescriptionLength
        - h1Count
        - headingCount
        - headings
        - visibleTextLength
        - linkCount
        - internalLinkCount
        - externalLinkCount
        - imageCount
        - imagesWithAlt
        - ctaCount
        - callsToAction
        - primaryCta
        - primaryCtaClearlyMarked
        - ctaAboveFold
        - contactSignals
        - legalSignals
        - companySignals
        - socialProofSignals
        - localSignals
        - technicalSignals
        - responseTimeMs
        - httpStatus
        - pageSizeKb
        - https
        - compressed
        - contentEncoding
        - canonical
        - robotsDirectives
        - openGraphSignals
        - hasViewport
        - hasLanguage
        - hasRobotsTxt
        - hasSitemap
        - sitemapUrl
        - localBusinessStructuredData
        - mapsLink
        - placeSignal
        - regionSignal
        - valuePropositionSignal
        - targetAudienceSignal
        - duplicateTextDetected
    ScanIssue:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        severity:
          type: string
          enum: [high, medium, low]
        impact:
          type: string
          enum: [high, medium, low]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        fact:
          type: string
        whyItMatters:
          type: string
        recommendation:
          type: string
        relatedChecks:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - severity
        - impact
        - difficulty
        - fact
        - whyItMatters
        - recommendation
        - relatedChecks
    AiRecommendation:
      type: object
      properties:
        title:
          type: string
        whatFound:
          type: string
        whyImportant:
          type: string
        whatToImprove:
          type: string
        proposal:
          type: ["string", "null"]
        impact:
          type: string
          enum: [high, medium, low]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        confidence:
          type: string
          enum: [high, medium, low]
        basedOnChecks:
          type: array
          items:
            type: string
      required:
        - title
        - whatFound
        - whyImportant
        - whatToImprove
        - proposal
        - impact
        - difficulty
        - confidence
        - basedOnChecks
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error