# lib/api-zod/src/generated/types/aiRecommendation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiRecommendationConfidence } from './aiRecommendationConfidence';
import type { AiRecommendationDifficulty } from './aiRecommendationDifficulty';
import type { AiRecommendationImpact } from './aiRecommendationImpact';

export interface AiRecommendation {
  title: string;
  whatFound: string;
  whyImportant: string;
  whatToImprove: string;
  /** @nullable */
  proposal: string | null;
  impact: AiRecommendationImpact;
  difficulty: AiRecommendationDifficulty;
  confidence: AiRecommendationConfidence;
  basedOnChecks: string[];
}