# lib/api-zod/src/generated/types/scanAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiRecommendation } from './aiRecommendation';
import type { CategoryScore } from './categoryScore';
import type { DetectedFacts } from './detectedFacts';
import type { ScanIssue } from './scanIssue';

export interface ScanAnalysis {
  /**
     * @minimum 0
     * @maximum 100
     */
  overallScore: number;
  aiRecommendations: AiRecommendation[] | null;
  categoryScores: CategoryScore[];
  detectedFacts: DetectedFacts;
  notChecked: string[];
  issues: ScanIssue[];
}